Replace long long with S64; replace unsigned long long with U64 change %lld to %"I64_FMT", etc. for porting support.
git-svn-id: file:///home/jj/hercules.svn/trunk@6906 956126f8-22a0-4046-8f4a-272fa8102e63

+++ ckddasd.c
@@ -958,7 +958,8 @@ ckd_read_track_retry:
 
     dev->syncio_active = active;
 
-    logdevtr (dev, MSG(HHC00429, "I", SSID_TO_LCSS(dev->ssid), dev->devnum, dev->filename, trk, f+1, (long long)dev->ckdtrkoff, dev->ckdtrksz));
+    logdevtr (dev, MSG(HHC00429, "I", SSID_TO_LCSS(dev->ssid), dev->devnum, dev->filename, trk, f+1, 
+        dev->ckdtrkoff, dev->ckdtrksz));
 
     /* Seek to the track image offset */
     offset = (off_t)dev->ckdtrkoff;
